refactor(sync): extract external-sync panel state into a hook
ExternalSyncPanel mixed all its fetch/mutation handlers with rendering
in one 380-line file. useExternalSyncPanel now owns mode switching,
form wiring and the sync/add/save/delete requests; the component is
left as a thin view over it, mirroring useCompareMode's split.
